requirement.txt
torch
torchvision
opencv-python
Pillow
numpy
supervision
ultralytics
scipy
pyyaml